Add home and privacy policy controllers to modularize route handling
